1. A brief History of Money
he history of currency and money could fill a full course or even a degree program.  In this first session, we will attempt to cover it in one class session.   Given this, by necessity, we will need to make significant simplifications and cover stretches of hundreds of years of monetary history in 1 session.
The objectives for this session are:
Understand the main functions of currency and be able to apply them when analyzing various types of historical, current or future currencies
Understand, at a high-level, the main forms of currencies used over time and the evolution toward fiat currency
Understand, from a monetary perspective, how Bitcoin, the most popular cryptocurrency, is designed
At some level, this is the most theoretical of the sessions as it is simply to provide a framework and some context about currency in general that might be helpful during the remainder of the course

https://courses.unic.ac.cy/course/view.php?id=5629#section-1
DFIN-511 Introduction to Digital Currencies  (Spring 2019- Registration Open)
The course is designed to provide an introductory understanding of decentralized digital currencies (cryptocurrencies) such as Bitcoin. In particular, the course will survey the theory and principles by which they operate, practical examples of basic transactions, the current and future interaction of cryptocurrencies with the banking, financial, legal and regulatory systems, how this ecosystem can be perceived as an innovation, and what role they may have for the developing world.


For students interested in continuing their studies, the University of Nicosia also offers the first Masters of Science Degree in Digital Currency, which is available for students worldwide, in the English language.

The live sessions run every Thursday or Friday, starting from February 1st onwards with world cryptocurrency experts Andreas Antonopoulos and Antonis Polemitis.
The content is created and updated by Andreas Vlachos, who also administrates the course.
The course runs from January 28th, 2019! Make sure to check the News Forum for recently posted important announcements!
Contacts:  digitalcurrency@unic.ac.cy  Twitter: @MScdigital @andr_vlachos
Conference Event Website: www.decentralized.com