Session 2 The Byzantine Generals’ Problem & the Bitcoin Solution
Objectives of Session 2
Thousands of cryptocurrencies/tokens exist out there. This session will provide the basis for understanding the theoretical underpinnings of Blockchain technology and Bitcoin – as the paradigm, thinking about their broader implications and the practical exercises that will follow in the later sessions. 
Students new to this technology will find this the most challenging session of the course.  It is important to invest as much time as is needed in understanding this part and to not be discouraged.  It took most of us many hours/days/weeks to understand this part! 
3
Recognize centralized asset & transaction ledgers  Understand the Byzantine Generals’ Problem Understand how Bitcoin addresses the Byzantine Generals’ Problem  Review some key Bitcoin metrics and developments in practice


DFIN-511: Introduction to Digital Currencies
https://courses.unic.ac.cy/pages/login.php
https://courses.unic.ac.cy/my/

|Thank You for Registering! Thank you for your interest in our free MOOC and MSc in Digital Currency at the University of Nicosia, world leader in cryptocurrency education.You have now been registered for our free introductory course DFIN-511: Introduction to Digital Currencies. We have sent you an email with instructions on how to login to our online learning platform and enroll to the course.