Session 3 Basics of Cryptocurrencies: Cryptography, transactions and mining
We are about to step into the more technical aspects of Bitcoin, as they are crucial in understanding how this system is enabled and how the Bitcoin network operates.  For the non-technical people, familiarization with several new concepts will be required, but we guarantee it will be a very rewarding experience in the long term!