Bitcoin clients,  wallets, cold storage, sending and receiving
Welcome to Introduction to Digital Currencies Free MOOC 11!  The course is designed to provide an introductory understanding of decentralized digital currencies (cryptocurrencies) such as Bitcoin. In particular, the course will survey the theory and principles by which they operate, practical examples of basic transactions, the current and future interaction of cryptocurrencies with the banking, financial, legal and regulatory systems, how this ecosystem can be perceived as an innovation, and what role they may have for the developing world. At the end of the Free MOOC, students who successfully qualify will receive their free academic certificates whose authenticity can be verified through the Bitcoin blockchain.