Session 5 Bitcoin in practice Part 2
Bitcoin Core, Constructing a transaction, Segwit, Lightning network, Mining pools    This is the last of the three introductory sessions on the technical aspects around the emerging field of cryptocurrencies. Exploring transactions over the basic Bitcoin client is the goal of this session, along with introducing scaling proposals such as SegWit and the Lightning Network. 
In the next sessions we’ll explore alternative currencies (altcoins) and alternate uses of the blockchain and then switch back to more business oriented subjects, under the view of cryptocurrenci